refactor: move js logic to stimulus
